src/main/java/com/example/service/UserService.java
package com.example.service;



import com.example.dao.LoginLogRepository;
import com.example.dao.UserRepository;
import com.example.domain.LoginLog;
import com.example.domain.User;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;


@Service
public class UserService {

    @Autowired
    private UserRepository userRepository;

    @Autowired
    private LoginLogRepository loginLogRepository;

    public boolean hasMatchUser(String userName, String password) {
         User findUser = userRepository.findUserByUserName(userName);
         if(findUser == null){
             return false;
         }
         if(findUser.getPassword().equals(password)){
             return true;
         }
        return false;
    }

    public User findUserByUserName(String userName) {
        return userRepository.findUserByUserName(userName);
    }

    public boolean saveLog(User user) {
        LoginLog loginLog = new LoginLog();
        loginLog.setUserId(user.getUserId());
        loginLog.setIp(user.getLastIp());
        loginLog.setLoginDate(user.getLastVisit());
        loginLogRepository.save(loginLog);
        return true;
    }
}
